change file operation method.

+++ Compress/file.h
@@ -1,14 +1,11 @@
-#include <string>  
-#include <fstream>
-#include <iostream>
 #include <vector>
 #include <map>
 using namespace std;
 
 namespace File
 {
-	int ReadFileToStringOrDie(const char* inputFilePath, string* data);
-	int WriteStringToFileOrDie(const string& data, const char* outputFilePath);
+	long ReadFileToStringOrDie(const char* inputFilePath, string* data);
+	long WriteStringToFileOrDie(const string& data, const char* outputFilePath);
 	vector<string> folderTraversalSingleLayerName(string dirPath);
 	map<string, int> folderTraversalSingleLayerNameSize(string dirPath);
 	void Txt2Binary(const char* inputFilePath, const char* outputFilePath);
